script/requirements.txt
playwright
psycopg2-binary
pg8000
python-dotenv
pytest
